+++ yogurt_editor/lib/src/cell_model.dart
@@ -5,7 +5,7 @@ import 'editor_controller.dart';
 abstract class CellModelBase {
   const CellModelBase();
 
-  List<CellPluginBase> get plugins => const [];
+  List<CellPlugin> get plugins => const [];
 
   Widget build(BuildContext context, CellState state);
 }
@@ -19,7 +19,7 @@ class CustomCellModel extends CellModelBase {
   final Widget Function(BuildContext context, CellState state) builder;
 
   @override
-  final List<CellPluginBase> plugins;
+  final List<CellPlugin> plugins;
 
   @override
   Widget build(BuildContext context, CellState state) {